% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syn.R
\name{addFertilisers}
\alias{addFertilisers}
\title{addFertilisers}
\usage{
addFertilisers(fertilisersAdded, fertiliserData)
}
\arguments{
\item{paddockHistory}{data frame containing paddock history.}
}
\description{
Calculate fertiliser N rate and net rate.
}
\author{
Fiona Evans
}
