% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syn.R
\name{calculateRON}
\alias{calculateRON}
\title{calculateRON}
\usage{
calculateRON(paddockHistory)
}
\arguments{
\item{paddockHistory}{data frame containing paddock history.}
}
\description{
calculate residual organic nitrogen (RON) estimate from paddock history.
}
\author{
Fiona Evans
}
