% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{color.of}
\alias{color.of}
\title{Makes colors for plotting}
\usage{
color.of(x, col = rainbow(100), range = NULL)
}
\arguments{
\item{x}{Input data.}

\item{col}{Color map (default is tim.colors).}

\item{range}{Range of x values to limit color to (not required).}
}
\value{
Returns a vector of colors.
}
\description{
Makes colors for plotting, stretched to cover the range of x
}
\author{
Fiona Evans
}
