% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syn.R
\name{fertiliserRateSensitivityAnalysis}
\alias{fertiliserRateSensitivityAnalysis}
\title{fertiliserRateSensitivityAnalysis}
\usage{
fertiliserRateSensitivityAnalysis(fert, crop, pyield, price, soil, oc,
  weeklyRain, fertilisersAdded, ron, fertiliserData, cropData,
  rates = seq(from = 0, to = 120, by = 12))
}
\arguments{
\item{fert}{xx}
}
\description{
fertiliserRateSensitivityAnalysis.
}
\author{
Fiona Evans
}
