% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syn.R
\name{getCropDataSYN}
\alias{getCropDataSYN}
\title{getCropDataSYN}
\usage{
getCropDataSYN()
}
\description{
Return crop data used by the SYN model. This data is used to calculate price adjustments 
for different protein / oil percentages in the grain.
}
\author{
Fiona Evans
}
