% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syn.R
\name{getRainfallProfilesSYN}
\alias{getRainfallProfilesSYN}
\title{getRainfallProfilesSYN}
\usage{
getRainfallProfilesSYN()
}
\description{
Return names of weekly rainfall profiles provided for default use by the SYN model.
}
\author{
Fiona Evans
}
