% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syn.R
\name{getWeeklyRainSYN}
\alias{getWeeklyRainSYN}
\title{getRainfallProfilesSYN}
\usage{
getWeeklyRainSYN(index)
}
\arguments{
\item{index}{Index of rainfall rpofile to use.}
}
\description{
Return data frame of weekly rainfall profiles used by the SYN model.
}
\author{
Fiona Evans
}
