% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{my.rescale}
\alias{my.rescale}
\title{Rescale a vector.}
\usage{
my.rescale(a, minb, maxb, mina = NULL, maxa = NULL)
}
\arguments{
\item{a}{Input vector.}

\item{minb}{.}

\item{maxb}{.}

\item{mina}{.}

\item{maxa}{.}
}
\description{
Linearly rescale a vector to range between minb and maxb.
}
\examples{
a <- c(2:8)
my.rescale(a, 1, 10)
my.rescale(a, 1, 10, minb=1, maxb=10)
}
\keyword{manip}
