% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npdecide.R
\name{NPdecide}
\alias{NPdecide}
\title{NPdecide}
\usage{
NPdecide(Nrate, Prate, Ypot, CfertN = 0.03, CstN = 0.05, Nst = 30,
  CfertP = 0.06, CstP = 0.12, Pst = 20, Area = NULL, P_y = NULL,
  P_nfert = NULL, P_pfert = NULL)
}
\arguments{
\item{Nrate}{N rate (kg/ha)}

\item{Prate}{P rate (kg/ha)}

\item{Ypot}{yield potential (kg/ha)}

\item{CfertN}{efficiency coefficient for applied N}

\item{CstN}{efficiency coefficient for soil N}

\item{Nst}{starting N in soil (kg/ha)}

\item{CfertP}{efficiency coefficient for applied P}

\item{CstP}{efficiency coefficient for soil P}

\item{Pst}{starting P in soil (kg/ha)}

\item{Area}{area sown (ha)}

\item{P_y}{price paid per kg of yield ($)}

\item{P_nfert}{price per kg of N fertiliser ($)}

\item{P_pfert}{price per kg of P fertiliser ($)}
}
\value{
Returns the predicted yield and net return if Area and prices are supplied.
}
\description{
Model for predicting effect on yield of aplications rates of nitrogen (N) and 
phosphorus (P) in the absence of any other yield-limiting factors.
}
\details{
The NPdecide model is described in detail in Robertson, M. J., Lyle, G and 
Bowden J. W. (2008), 
Within-field variability of wheat yield and economic implications for spatially variable 
nutrient management, Field Crops Research Vol 105(3), pp 211-220.

Input parameters CstP, cFertN, Cst, cfertN, Nst and Pst relate to amoutn of fertilisers already
available int he soil, and the responsiveness of the soil to added fertilisers. Initial values
have been set according to the references paper, but would normally be available from soil tests.


The model is also used in an online tool: http://optimiser.csiro.au/
}
\examples{
# Plot response curves

Nrate <- seq(from=0, to=200, by=20) 
Nyield <- rep(NA, length(Nrate))
for (i in 1:length(Nrate)) {
  Nyield[i] <- NPdecide(Nrate[i], 150, Ypot=3000)$yield
}

Prate <- seq(from=0, to=30, by=2) 
Pyield <- rep(NA, length(Prate))
for (i in 1:length(Prate)) {
  Pyield[i] <- NPdecide(500, Prate[i], Ypot=3000)$yield
}

par(mfrow=c(1,2)) 
plot(Nrate, Nyield, type="b", main="Yield response to N (P non-limiting)", xlab="Yield (kg/ha)") 
plot(Prate, Pyield, type="b", main="Yield response to P (N non-limiting)", xlab="Yield (kg/ha)") 

}
\author{
Fiona Evans
}
