% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npoptim.R
\name{NPoptim}
\alias{NPoptim}
\title{Optimise rates of nitrogen (N) and phosphorus (P)}
\usage{
NPoptim(c(90,60,0,32,23,20), yfunc3, control=list(fnscale=-1))
}
\arguments{
\item{rate}{vector containing starting values for N rate (first 3) and then  P rate (last 3).}
}
\value{
Returns a vector of station ids, in order of increasing distance.
}
\description{
Optimise rates of nitrogen (N) and phosphorus (P) 
over 3 paddocks or management zones. This function is used as the base model in an 
online fertiliser tool for precision agriculture that is available at http://optimiser.csiro.au
}
\author{
Fiona Evans
}
