% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syn.R
\name{SYN}
\alias{SYN}
\title{SYN}
\usage{
SYN(currentCrop, currentPotentialYield, currentCropBasePrice, soil,
  soilOrganicCarbon, weeklyRain, fertilisersAdded, RONcarried, fertiliserData,
  cropData)
}
\arguments{
\item{paddockHistory}{data frame containing paddock history.}
}
\description{
Run the SYN model.
}
\author{
Fiona Evans
}
