% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syn.R
\name{SYNsensitivity}
\alias{SYNsensitivity}
\title{SYNsensitivity}
\usage{
SYNsensitivity(currentCrop, currentPotentialYield, currentCropBasePrice, TON,
  fertilisersAdded, fertiliserData, cropData, parameters)
}
\arguments{
\item{currentCrop}{element of c( "ASW", "Premium White wheat", "Hard wheat", "Durum Premium", "Durum General",}

\item{currentPotentialYield}{numeric (t/ha)}

\item{currentCropBasePrice}{numeric ($)}

\item{TON}{Total Organic Nitrogen object}

\item{fertilisersAdded}{based on fertilisers returned by getFertiliserDataSYN, see examples for format}

\item{fertiliserData}{Fertiliser data and costs, an example can be obtained using getFertiliserDataSYN()}

\item{cropData}{Data for protein premium calculations, an example can be obtained using getCropDataSYN()}

\item{parameters}{Parameters used in yield and protein calculations, these are constants in the SYN function.}
}
\description{
Run the SYN model without the weekly timestep soil water, root growth and leaching part.
}
\author{
Fiona Evans
}
