% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{two.colors.old}
\alias{two.colors.old}
\title{Older version from package 'fields' (less of the middle colour present).}
\usage{
two.colors.old(n = 256, start = "darkgreen", end = "red",
  middle = "white")
}
\arguments{
\item{n}{Length of output vector.}

\item{start}{Color.}

\item{end}{Color.}

\item{middle}{Color.}
}
\description{
Color interpolation between three colors to output a color vector.
}
\keyword{color}
