% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syn.R
\name{fertiliserRateSensitivityAnalysis}
\alias{fertiliserRateSensitivityAnalysis}
\title{fertiliserRateSensitivityAnalysis}
\usage{
fertiliserRateSensitivityAnalysis(fert, crop, pyield, price, soil, oc,
  weeklyRain, fertilisersAdded, ron)
}
\arguments{
\item{fert}{xx}
}
\description{
fertiliserRateSensitivityAnalysis.
}
\author{
Fiona Evans
}
