% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syn.R
\name{SYN}
\alias{SYN}
\title{SYN}
\usage{
SYN(currentCrop, currentPotentialYield, currentCropBasePrice, soil,
  soilOrganicCarbon, weeklyRain, fertilisersAdded, RONcarried, fertiliserData,
  cropData)
}
\arguments{
\item{currentCrop}{element of c( "ASW", "Premium White wheat", "Hard wheat", "Durum Premium", "Durum General",}

\item{currentPotentialYield}{numeric (t/ha)}

\item{currentCropBasePrice}{numrix ($)}

\item{soil}{an element of soils, use function getSoilsSYN()}

\item{soilOrganicCarbon}{percentage (between 0 and 100)}

\item{weeklyRain}{weeklyRain is a rainfall profile, use getRainfallProfilesSYN}

\item{fertilisersAdded}{based on fertilisers returned by getFertiliserDataSYN, see examples for format}

\item{RONcarried}{numeric, can be calculated from paddock history.}
}
\description{
Run the SYN model. This version of the SYN model has been encoded from the original Excel spreadsheet 
version of the model, available from:
}
\examples{
# Read model documentation for the original spreadsheet model:
f <- system.file("doc", "Select Your Nitrogen Users Manual.pdf", package = "Ragronomy")
shell.exec(normalizePath(f))

# Set model inputs
currentCrop <- "ASW"

currentPotentialYield <- 2

currentCropBasePrice <- 202

soils <- getSoilsSYN()
soil <- soils[1,]

soilOrganicCarbon <- 0.5

getRainfallProfilesSYN()
weeklyRain <- getWeeklyRainSYN(5)

fertilisersAdded <- data.frame(name=c("DAP", "Urea", "nitrate"), 
                               weekApplied=c(0, 4, -14), 
                               netRate=c(15, 30, 50))

RONcarried <- calculateRON(paddockHistory)

fertiliserData <- getFertiliserDataSYN()
cropData <- getCropDataSYN()

# Run the model
SYN(currentCrop, currentPotentialYield, currentCropBasePrice,
    soil, soilOrganicCarbon, weeklyRain, fertilisersAdded, 
   RONcarried, fertiliserData, cropData)

}
\author{
Fiona Evans
}
